Implemented voiced navigation
Implements #22

Implemented text to speech for the 4 arrows. When turning left, the navigator will say "turning left" and so on. Added the necessary method to play mp3 files and added the duration as a parameter too. The navigator says "successfully connected to your car when an mqtt connection is established". Will add an option to toggle it off and on later.