Dark mode and interface
Fixed some more of darkmode and removed some unused/unneeded code. There's still a persistent bug where it send you bakc to HomeFragment once you activate it. Also implemented the log in screen and a button to switch to and from it, the button wouldn't be there in the final product but due to time restraints it will have to stay for this project.